+++ Vector/vector.h
@@ -4,19 +4,19 @@
 #include <stdlib.h>
 
 typedef struct {
-    void **data;     /* information stored in the vector */
-    size_t count;    /* number of elements currently used in the vector */
-    size_t capacity; /* maximum capacity of the vector */
+    void **data;           /* information stored in the vector */
+    unsigned int count;    /* number of elements currently used in the vector */
+    unsigned int capacity; /* maximum capacity of the vector */
 } Vector;
 
-extern const size_t VECTOR_INIT_CAPACITY;
+extern const unsigned int VECTOR_INIT_CAPACITY;
 
 Vector *vector_create();
 void vector_free(Vector *vector);
 
 int vector_add(Vector *vector, void *item);
-int vector_insert(Vector *vector, void *item, size_t index);
-int vector_delete(Vector *vector, size_t index);
+int vector_insert(Vector *vector, void *item, unsigned int index);
+int vector_delete(Vector *vector, unsigned int index);
 
 
 #endif // CDSLIB_VECTOR_H
